split the attribute_values per type
There is really no use-case where anyone is looking for an attribute
value, but doesn't care for what type of object. Rather, the user will
know in advance if they're looking for a signal attribute, a message
attribute, etc.

Splitting the attributes by type will simplify everyone's code, since a
match to extract only the useful values will no longer be needed.

Additionally, this makes searching for a specific attribute more
efficient, since the list to consider will be smaller.